+++ src/main/java/parser/Parser.java
@@ -1,16 +1,18 @@
 package parser;
 
+import command.*;
+
 public class Parser {
     private final int lengthOfSignature = 3;
-    public command parseCommand(String input){
+    public Command parseCommand(String input){
         input = input.trim();
 
         if(!input.contains(" ")){
-            if(input == "bye"){
-                return new ExitCommand();
+            if(input.equals("bye")){
+                return new EndCommand();
             }
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
 
         int indexOfSpace = input.indexOf(" ");
@@ -20,29 +22,25 @@ public class Parser {
         switch(type) {
         case "appointment":
             return parseAppointment(statement);
-            break;
-        case "task":
-            return parseTask(statement);
-            break;
         case "pet":
             return parsePet(statement);
-            break;
-        case "staff":
-            return parseStaff(statement);
-            break;
+        case "employee":
+            return parseEmployee(statement);
+        case "service":
+            return parseService(statement);
         default:
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
     }
 
-    public command parseAppointment(String input){
+    public Command parseAppointment(String input){
         if(!input.contains(" ")){
-            if(input == "view"){
-                return new AppointmentViewCommand();
+            if(input.equals("view")){
+                return new ViewAppointmentCommand();
             }
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
 
         String type = input.substring(0,input.indexOf(" "));
@@ -54,19 +52,95 @@ public class Parser {
         case RemoveAppointmentCommand.COMMAND_WORD:
             return prepareRemoveAppointment(statement);
         break;
-        case ViewAppointmentCommand.COMMAND_WORD:
-            return new ViewAppointmentCommand();
-        break;
         case AllocateAppointmentCommand.COMMAND_WORD:
             return prepareAllocateAppointment(statement);
         break;
+        case SetAppointmentStatusCommand.COMMAND_WORD:
+            return prepareSetAppointmentStatusCommand(statement);
+            break;
         default:
-            return new ExitCommand();
+            return new EndCommand();
         }
     }
 
-    public command prepareAllocateAppointment(String input){
 
+    public Command parseService(String input){
+        if(!input.contains(" ")){
+            if(input.equals("view")){
+                return new ViewServiceCommand();
+            }
+            System.out.println("input invalid");
+            return new EndCommand();
+        }
+
+        String type = input.substring(0,input.indexOf(" "));
+        String statement = input.substring(input.indexOf(" ")).trim();
+        switch(type) {
+        case AddServiceCommand.COMMAND_WORD:
+            return prepareAddService(statement);
+        break;
+        case RemoveServiceCommand.COMMAND_WORD:
+            return prepareRemoveService(statement);
+        break;
+        default:
+            return new EndCommand();
+        }
+    }
+
+    public Command prepareAddService(String input){
+        int d = input.indexOf(" d/");
+
+        if(d == -1){
+            System.out.println("invalid input");
+            return new EndCommand();
+        }
+
+        String description = input.substring(d + lengthOfSignature);
+
+        return new AddServiceCommand(description);
+
+    }
+
+
+    public Command prepareRemoveService(String input){
+        int index = indexOfRemove(input);
+        if(index == -1){
+            System.out.println("input invalid");
+            return new EndCommand();
+        }
+
+        return new RemoveServiceCommand(index);
+    }
+
+    public Command prepareSetAppointmentStatusCommand(String input){
+        int i = input.indexOf(" i/");
+        int s = input.indexOf(" s/");
+
+        if(i > s || i == -1 || s == -1){
+            System.out.println("invalid input");
+            return new EndCommand();
+        }
+
+        String index = input.substring(i + lengthOfSignature, s);
+        String status = input.substring(s + lengthOfSignature);
+
+        return new AllocateApointmentCommand(index, status);
+    }
+
+
+    public Command prepareAllocateAppointment(String input){
+        int i = input.indexOf(" i/");
+        int n = input.indexOf(" n/");
+
+        if(i > n || i == -1 || n == -1){
+            System.out.println("invalid input");
+            return new EndCommand();
+        }
+
+        String index = input.substring(i + lengthOfSignature, n);
+        String name = input.substring(n + lengthOfSignature);
+
+        return new AllocateApointmentCommand(index, name);
     }
 
     private int numOfSpace(String input){
@@ -81,21 +155,21 @@ public class Parser {
     }
 
 
-    public command prepareRemoveAppointment(String input){
+    public Command prepareRemoveAppointment(String input){
         int index = indexOfRemove(input);
         if(index == -1){
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
 
         return new RemoveAppointmentCommand(index);
     }
 
-    public command prepareRemovePet(String input){
+    public Command prepareRemovePet(String input){
         int index = indexOfRemove(input);
         if(index == -1){
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
 
         return new RemovePetCommand(index);
@@ -111,19 +185,18 @@ public class Parser {
             return -1;
         }
 
-        int appointmentId = Integer.parseInt(id);
-        return appointmentId;
+        return Integer.parseInt(id);
     }
 
 
-    public command prepareAddAppointment(String input){
+    public Command prepareAddAppointment(String input){
         int s = input.indexOf(" s/");
         int p = input.indexOf(" p/");
         int d = input.indexOf(" d/");
 
         if(s > p || p > d || s == -1 || p == -1 || d == -1){
             System.out.println("invalid input");
-            return new ExitCommand();
+            return new EndCommand();
         }
 
         String service = input.substring(s + lengthOfSignature, p);
@@ -135,23 +208,21 @@ public class Parser {
 
     public boolean isInt(String val){
         Boolean strResult = val.matches("\\d?");
-        if(strResult == true) {
+        if(strResult) {
             return true;
         }
         return false;
     }
 
 
-    public command parseTask(String input){
 
-    }
-    public command prepareAddPet(String input){
+    public Command prepareAddPet(String input){
         int startOfN = input.indexOf(" n/");
         int startOfS = input.indexOf(" s/");
 
         if(startOfN > startOfS || startOfN == -1|| startOfS == -1){
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
 
         String name = input.substring(startOfN + lengthOfSignature, startOfS);
@@ -160,13 +231,13 @@ public class Parser {
     }
 
 
-    public command parsePet(String input){
+    public Command parsePet(String input){
         if(!input.contains(" ")){
-            if(input == "view"){
+            if(input.equals("view")){
                 return new AppointmentViewCommand();
             }
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
 
         String type = input.substring(0,input.indexOf(" "));
@@ -174,17 +245,67 @@ public class Parser {
         switch(type) {
         case AddPetCommand.COMMAND_WORD:
             return prepareAddPet(statement);
-        break;
+            break;
         case RemoveAppointmentCommand.COMMAND_WORD:
             return prepareRemovePet(statement);
-        break;
+            break;
         case ListPetsCommand.COMMAND_WORD:
             return new ListPetsCommand();
             break;
+        case AllocateAppointmentCommand.COMMAND_WORD:
+            return prepareAllocateAppointment(statement);
+            break;
         default:
             System.out.println("input invalid");
-            return new ExitCommand();
+            return new EndCommand();
         }
+    }
+
+    public Command parseEmployee(String input){
+        if(!input.contains(" ")){
+            if(input.equals("view")){
+                return new ViewEmployeeCommand();
+            }
+            System.out.println("input invalid");
+            return new EndCommand();
+        }
+
+        String type = input.substring(0,input.indexOf(" "));
+        String statement = input.substring(input.indexOf(" ")).trim();
+        switch(type) {
+        case AddEmployeeCommand.COMMAND_WORD:
+            return prepareAddEmployee(statement);
+            break;
+        case RemoveEmployeeCommand.COMMAND_WORD:
+            return prepareRemoveEmployee(statement);
+            break;
+        default:
+            System.out.println("input invalid");
+            return new EndCommand();
+        }
+    }
+
+    public Command prepareAddEmployee(String input){
+        int startOfN = input.indexOf(" n/");
+
+        if(startOfN == -1){
+            System.out.println("input invalid");
+            return new EndCommand();
+        }
+
+        String name = input.substring(startOfN + lengthOfSignature);
+        return new AddEmployeeCommand(name);
+    }
+
+
+    public Command prepareRemoveEmployee(String input){
+        int index = indexOfRemove(input);
+        if(index == -1){
+            System.out.println("input invalid");
+            return new EndCommand();
+        }
+
+        return new RemoveEmployeeCommand(index);
     }
